requirements-minimal.txt
flask
langchain
langchain-openai
python-dotenv
openai
